Update AreaOfRectangle.c
There is one thing that could be improved in the code. The code does not check if the user enters valid values for the length and breadth of the rectangle. For example, if the user enters a negative value for the length or breadth, the code will produce an incorrect result.