Update LABESE1.c
In the sale[i] = round(arr[i]*0.9) - 0.01 line, this code is rounding the result of the multiplication to the nearest pound before subtracting the 1 penny. This means that if the original price is a multiple of 100, the sale price will always be less than the original price. If the sale price is negative, then the calculation is invalid and you should output an error message.