Update arm.c
The problem with this code is that the pow function in the Finding Sum loop is not correctly calculating the power of the individual digits. The pow function takes two arguments: the base and the exponent. In this code, you are passing the digit itself as both the base and the exponent. This is incorrect because the exponent should be the number of digits in the input number.