Update Count_Index.py
I created a new function, find_positions(), to encapsulate the logic for finding the positions of an element in a list. This makes the code more modular and reusable.
I used a while loop instead of a for loop to find the positions of x in the list. This is more efficient because it avoids having to iterate over the list multiple times.
I used a list.remove() instead of list.pop() to remove x from the list. This is because list. remove() is more efficient for removing elements from the middle of a list.
I added a check to see if the positions list is empty before printing the results. This prevents the code from printing "This Number is present at [] place(s)---[index given]" if x is not in the list.